Fix issue with exponential envelope with non-zero delay
In a corner case, where the state becomes STATE_IDLE EXACTLY at the end of an update(), AND noteOn() is called before the next update() occurs, AND delay_count is non-zero, AND the exponential envelope shape is in use, exp_count will have a stale value of 2^32-1 and the delay becomes about 9 days. Some code duplicated here - it might be possible to move it instead.